requirements.txt
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-debug-toolbar==2.2
django-redis==4.11.0
idna==2.9
mysqlclient==1.4.6
Pillow==7.0.0
PyMySQL==0.9.3
pytz==2019.3
redis==3.4.1
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.8
xlwt==1.3.0
